Move legacy contracts out of contracts folder and their build artefacts out of the default /build folder
