+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 build
+build-legacy
 tmp
 bin
 .outputParameter
