+++ contracts-legacy/GuardianStorage.sol
