+++ contracts-legacy/IGuardianStorage.sol
