+++ contracts-legacy/IUniswapExchange.sol
