+++ test/integration/makerV2Manager_loan.js
@@ -18,7 +18,7 @@ const Join = require("../../build/JoinLike");
 const Vat = require("../../build/VatLike");
 const FaucetUser = require("../../build/FaucetUser");
 const CdpManager = require("../../build/TestCdpManager");
-const MakerV1Manager = require("../../build/LegacyMakerManager");
+const MakerV1Manager = require("../../build-legacy/LegacyMakerManager");
 const MakerV2Manager = require("../../build/MakerV2Manager");
 const UpgradedMakerV2Manager = require("../../build/TestUpgradedMakerV2Manager");
 const TransferManager = require("../../build/TransferManager");
