+++ test/makerV2Manager_loan.js
@@ -10,7 +10,7 @@ const { HashZero, AddressZero } = ethers.constants;
 const TestManager = require("../utils/test-manager");
 const GemJoin = require("../build/GemJoin");
 const Registry = require("../build/ModuleRegistry");
-const MakerV1Manager = require("../build/LegacyMakerManager");
+const MakerV1Manager = require("../build-legacy/LegacyMakerManager");
 const MakerV2Manager = require("../build/MakerV2Manager");
 const UpgradedMakerV2Manager = require("../build/TestUpgradedMakerV2Manager");
 const MakerRegistry = require("../build/MakerRegistry");
