+++ test/transferManager.js
@@ -6,7 +6,7 @@ const Registry = require("../build/ModuleRegistry");
 const TransferStorage = require("../build/TransferStorage");
 const GuardianStorage = require("../build/GuardianStorage");
 const TransferModule = require("../build/TransferManager");
-const LegacyTransferManager = require("../build/LegacyTransferManager");
+const LegacyTransferManager = require("../build-legacy/LegacyTransferManager");
 const TokenPriceProvider = require("../build/TokenPriceProvider");
 const ERC20 = require("../build/TestERC20");
 const TestContract = require("../build/TestContract");
